requirements.txt

streamlit
PyPDF2
python-pptx
docx2txt
nltk
scikit-learn
textstat
plotly
python-docx
supabase
python-magic-bin
